llama_core

Module search

Available on crate feature search only.

Define APIs for web search operations.

Structs

SearchConfig

The base Search Configuration holding all relevant information to access a search api and retrieve results.

SearchOutput

Final output format for consumption by the LLM.

SearchResult

output format for individual results in the final output.

Enums

ContentType

Possible input/output Content Types. Currently only supports JSON.