Crate llama_core

Llama Core, abbreviated as llama-core, defines a set of APIs. Developers can utilize these APIs to build applications based on large models, such as chatbots, RAG, and more.

Re-exports

pub use error::LlamaCoreError;

pub use graph::EngineType;

pub use graph::Graph;

pub use graph::GraphBuilder;

pub use metadata::ggml::GgmlMetadata;

pub use metadata::piper::PiperMetadata;

pub use metadata::whisper::WhisperMetadata;

pub use metadata::BaseMetadata;

Modules

audio

Define APIs for audio generation, transcription, and translation.

chat

Define APIs for chat completion.

completions

Define APIs for completions.

embeddings

Define APIs for computing embeddings.

error

Error types for the Llama Core library.

graph

Define Graph and GraphBuilder APIs for creating a new computation graph.

images

Define APIs for image generation and edit.

metadata

models

Define APIs for querying models.

rag

Define APIs for RAG operations.

search

utils

Define utility functions.

Structs

PluginInfo

Version info of the wasi-nn_ggml plugin, including the build number and the commit id.

Enums

RunningMode

Running mode

SDContextType

The context to create for the stable diffusion model

Functions

get_plugin_info

Get the plugin info

init_core_context

Initialize the core context

init_piper_context

Initialize the piper context

init_rag_core_context

Initialize the core context for RAG scenarios.

init_sd_context_with_full_model

Initialize the stable diffusion context with the given full diffusion model

init_sd_context_with_standalone_model

Initialize the stable diffusion context with the given standalone diffusion model

init_whisper_context

Initialize the whisper context

running_mode

Return the current running mode.