Crate llama_core

Llama Core, abbreviated as llama-core, defines a set of APIs. Developers can utilize these APIs to build applications based on large models, such as chatbots, RAG, and more.

Re-exports

pub use error::LlamaCoreError;

Modules

chat

Define APIs for chat completion.

completions

Define APIs for completions.

embeddings

Define APIs for computing embeddings.

error

Error types for the Llama Core library.

models

Define APIs for querying models.

rag

Define APIs for RAG operations.

utils

Define utility functions.

Structs

Graph

Wrapper of the wasmedge_wasi_nn::Graph struct

Metadata

Model metadata

MetadataBuilder

Builder for the Metadata struct

PluginInfo

Version info of the wasi-nn_ggml plugin, including the build number and the commit id.

Enums

RunningMode

Running mode

Functions

get_plugin_info

Get the plugin info

init_core_context

Initialize the core context

init_rag_core_context

Initialize the core context for RAG scenarios.

running_mode

Return the current running mode.